+++ dotnet/Data/MerchantSettingsRepository.cs
@@ -64,7 +64,7 @@
                 merchantSettings.ItemModals.Add(new ModalMap("WHITE_GOODS", "NONE", false));
                 merchantSettings.ItemModals.Add(new ModalMap("FIREARMS", "ORM_D", false));
             }
-            if (merchantSettings.SlaSettings.Count != 14)
+            if (merchantSettings.SlaSettings.Count != 19)
             {
                 merchantSettings.SlaSettings = new List<SlaSettings>();
                 merchantSettings.SlaSettings.Add(new SlaSettings("FedEx Ground", false, 0, 0));
@@ -75,12 +75,17 @@
                 merchantSettings.SlaSettings.Add(new SlaSettings("Standard Overnight", false, 0, 0));
                 merchantSettings.SlaSettings.Add(new SlaSettings("2Day", false, 0, 0));
                 merchantSettings.SlaSettings.Add(new SlaSettings("FedEx Home Delivery", false, 0, 0));
+                merchantSettings.SlaSettings.Add(new SlaSettings("First Overnight Freight", false, 0, 0));
+                merchantSettings.SlaSettings.Add(new SlaSettings("1 Day Freight", false, 0, 0));
+                merchantSettings.SlaSettings.Add(new SlaSettings("2 Day Freight", false, 0, 0));
+                merchantSettings.SlaSettings.Add(new SlaSettings("3 Day Freight", false, 0, 0));
                 merchantSettings.SlaSettings.Add(new SlaSettings("International Economy", false, 0, 0));
                 merchantSettings.SlaSettings.Add(new SlaSettings("International Priority Express (IP EXP)", false, 0, 0));
                 merchantSettings.SlaSettings.Add(new SlaSettings("International Priority EOD (IP EOD)", false, 0, 0));
                 merchantSettings.SlaSettings.Add(new SlaSettings("International Connect Plus", false, 0, 0));
                 merchantSettings.SlaSettings.Add(new SlaSettings("International First", false, 0, 0));
                 merchantSettings.SlaSettings.Add(new SlaSettings("International Two Day", false, 0, 0));
+                merchantSettings.SlaSettings.Add(new SlaSettings("International Priority Freight", false, 0, 0));
             }
             if (merchantSettings.PackingAccessKey == null) 
             {